refactor: adapt test imports
