+++ src/core/tests/notifyManager.test.tsx
@@ -1,5 +1,5 @@
 import { createNotifyManager } from '../notifyManager'
-import { sleep } from '../../devtools/tests/utils'
+import { sleep } from '../../tests/utils'
 
 describe('notifyManager', () => {
   it('should use default notifyFn', async () => {
