+++ src/createAsyncStoragePersister/tests/asyncThrottle.test.ts
@@ -1,5 +1,5 @@
 import { asyncThrottle } from '../asyncThrottle'
-import { sleep as delay } from '../../reactjs/tests/utils'
+import { sleep as delay } from '../../tests/utils'
 
 describe('asyncThrottle', () => {
   test('basic', async () => {
