+++ src/createWebStoragePersister/tests/storageIsFull.test.ts
@@ -1,6 +1,6 @@
 import { dehydrate, MutationCache, QueryCache, QueryClient } from '../../core'
 import { createWebStoragePersister } from '../index'
-import { sleep } from '../../reactjs/tests/utils'
+import { sleep } from '../../tests/utils'
 
 function getMockStorage(limitSize?: number) {
   const dataSet = new Map<string, string>()
