+++ src/reactjs/tests/QueryClientProvider.test.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { render, waitFor } from '@testing-library/react'
 import { renderToString } from 'react-dom/server'
 
-import { sleep, queryKey, createQueryClient } from './utils'
+import { sleep, queryKey, createQueryClient } from '../../tests/utils'
 import {
   QueryClient,
   QueryClientProvider,
