+++ src/reactjs/tests/QueryResetErrorBoundary.test.tsx
@@ -2,7 +2,8 @@ import { waitFor, fireEvent } from '@testing-library/react'
 import { ErrorBoundary } from 'react-error-boundary'
 import React from 'react'
 
-import { sleep, queryKey, renderWithClient, createQueryClient } from './utils'
+import { sleep, queryKey, createQueryClient } from '../../tests/utils'
+import { renderWithClient } from './utils'
 import { useQuery, QueryCache, QueryErrorResetBoundary } from '../..'
 
 describe('QueryErrorResetBoundary', () => {
